requirements.txt
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.3
requests==2.31.0
tomli==2.0.1
urllib3==2.1.0
